feat: added crud on orders and handled bad request

+++ src/Routes/user.ts
@@ -1,49 +1,53 @@
-import {User} from "@prisma/client";
 import prisma from "../Config/Database";
 import express from "express";
+import {safeExecutionWrapper} from "../Helper/ErrorCatcher";
+import {WishDto} from "../Helper/WishDto";
+import {userConnectedDto, userUpdateDto} from "../dto/userDto";
 
 const router = express.Router();
 
-router.get("/", async (req, res, next) => {
-    try {
-        const users: User[] = await prisma.user.findMany()
-        res.status(200).json(users);
-    } catch (e: any) {
-        next(e);
-    }
-});
-
-router.put("/:id", async (req, res, next) => {
-    try {
-        const {id} = req.params;
-        const {firstname, lastname} = req.body;
-        const user = await prisma.user.update({
+router.put("/update", async (req, res, next) => {
+    await safeExecutionWrapper(res, async () => {
+        if (!WishDto(userUpdateDto, req.body, res)) return;
+        const {firstname, lastname, user} = req.body;
+        const userData = await prisma.user.update({
             where: {
-                id: id,
+                id: user.id,
             },
             data: {
                 firstName: firstname,
                 lastName: lastname,
+            },
+            select: {
+                id: true,
+                firstName: true,
+                lastName: true,
+                email: true,
+                role: true,
             }
         });
-        res.status(200).json(user);
-    } catch (e: any) {
-        next(e);
-    }
+        res.status(200).json(userData);
+    });
 });
 
-router.delete("/:id", async (req, res, next) => {
-    try {
-        const {id} = req.params;
+router.delete("/delete", async (req, res, next) => {
+    await safeExecutionWrapper(res, async () => {
+        if (!WishDto(userConnectedDto, req.body, res)) return;
+        const {id} = req.body.user;
         const user = await prisma.user.delete({
             where: {
                 id: id,
+            },
+            select: {
+                id: true,
+                firstName: true,
+                lastName: true,
+                email: true,
+                role: true,
             }
         });
         res.status(200).json(user);
-    } catch (e: any) {
-        next(e);
-    }
+    });
 });
 
 export default router;
